+++ infra_10x/namespace.py
@@ -93,7 +93,7 @@ class CUSTOM_ROLE(NamedConstant):
 
 
 # fmt: off
-SECURITY_KEYS_MISSING = 'Sec keys are missing. If you are using a new computer, please run "xx user new machine" from your shell'
+SECURITY_KEYS_MISSING = 'Sec keys are missing. If you are using a new computer, please run: xx-user-init --new-machine'
 SECURITY_KEYS_INCOMPATIBLE = 'Sec keys are incompatible with an encrypted password'
 
 SECRETS_CLIENT_ARGS     = {'service_name': 'secretmanager', 'region_name': 'me-south-1'}